♻️ refactor: ChapterSelector 공통 컴포넌트 경로  이동
