+++ src/shared/ui/segment/ChapterSelector.tsx
